Support draining storage
This pivots to a situation where the calling code explicitly chooses
what lifecycle state the log is in when it constructs it. This means
that a log constructed in the draining state can explicitly not have any
writes made to it, but it has the required database parameters still,
and thus can support integration of any pending leaves that were
previously sequenced.

On paper this approach seemed a bit overkill, but it actually feels
quite elegant when written.